refactor(lab/14): move all the simulation stuff into it's own module
add some notes about use cases it should support
